EmployeeDataRepository - turn info an interface
